refactoring: restructured architecture
