+++ src/components/About.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+class About extends React.Component {
+  render() {
+    return (
+      <>
+        <h1>Привет! Меня зовут Слава</h1>
+      </>
+    );
+  }
+}
+
+export default About;
