+++ src/components/Cards.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+class Cards extends React.Component {
+  render() {
+    return (
+      <>
+        <div className="cards">Cards</div>
+      </>
+    );
+  }
+}
+
+export default Cards;
