+++ src/main.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import { createBrowserRouter, RouterProvider } from 'react-router-dom';
-import AboutPage from './pages/about/AboutPage';
-import CardsPage from './pages/cards/CardsPage';
+import AboutPage from './pages/AboutPage';
+import CardsPage from './pages/CardsPage';
 
 const router = createBrowserRouter([
   {
